tools/arch-review/requirements.txt
anthropic==0.40.0
PyGithub==2.1.1
GitPython==3.1.40
